Modularized cart list and cart view component
