+++ ShopHub/ShopHub/User Interface/Page Views/Cart/CartView.swift
@@ -9,9 +9,6 @@ import SwiftUI
 
 struct CartView: View {
     
-    // Environment Object
-    @EnvironmentObject var cart: CartViewModel
-    
     // Internal State
     @State private var searchText: String = ""
     @State private var isLogoPressed = false
@@ -19,16 +16,7 @@ struct CartView: View {
     var body: some View {
         NavigationStack {
             VStack {
-                // MARK: Product list in users' cart
-                List {
-                    ForEach(cart.sectionHeaders, id: \.self) { type in
-                        Section(header: Text(cart.sectionHeader(type)).font(.subheadline)) {
-                            ForEach(cart.sectionContent(type)) { product in
-                                ProductCardView(product: product)
-                            }
-                        }
-                    }
-                }
+                CartListView()
             }
             .toolBarStyle(title: "Shopping Cart", titleImage: "cart.fill", isLogoPressed: $isLogoPressed)
             .searchable(text: $searchText, prompt: "Search for your product in cart")
@@ -38,33 +26,6 @@ struct CartView: View {
 
 #Preview {
     CartView()
+        .environmentObject(CartViewModel())
 }
 
-// TODO: Design UI
-struct ProductCardView: View {
-    let product: Product
-    var body: some View {
-        HStack(alignment: .center, spacing: 10) {
-            Image(product.image)
-                .resizable()
-                .scaledToFit()
-                .frame(width: 100, height: 100)
-                .padding(10)
-            
-            VStack(alignment: .leading, spacing: 5) {
-                Text(product.name)
-                    .font(.headline)
-
-                Text(product.description ?? "N/A")
-                    .font(.subheadline)
-                    .lineLimit(2)
-            }
-        }
-    }
-}
-
-#Preview("Product Card") {
-    let products: [Product] = Bundle.main.decode("ProductList.json")
-
-    return ProductCardView(product: products[1])
-}
